% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{aggregateMatches}
\alias{aggregateMatches}
\title{aggregateMatches}
\usage{
aggregateMatches(e, fn = agg.repr)
}
\arguments{
\item{e}{A GRanges object as produced by `findSeedMatches`.}
}
\value{
An aggregated data.frame
}
\description{
aggregateMatches
}
