% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{aggregateMatches_Biochem}
\alias{aggregateMatches_Biochem}
\title{aggregateMatches_Biochem}
\usage{
aggregateMatches_Biochem(
  e,
  kd_cut_off = 0,
  ag = -6.5,
  ORF = TRUE,
  keepSiteInfo = FALSE
)
}
\arguments{
\item{e}{A GRanges object as produced by `findSeedMatches`.}

\item{kd_cut_off}{A cutoff value for log_kd values}

\item{ag}{The 'ag' value for the aggregation, corresponding to the free
concentration of AGO}

\item{ORF}{Option indicating whether sites in the open reading frame are
included in the scan. If they are included, a column named 'ORF' with
'TRUE' / 'FALSE' entries should characterize each site.}

\item{keepSiteInfo}{An option on wether to also keep info concerning the
number of Binding Sites}
}
\value{
An aggregated data.frame
}
\description{
Aggregates Matches of the findSeedMatches function according to the
"Biochemical Model" of McGeary et al., 2020, Science
}
