% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregateSites}
\alias{aggregateSites}
\title{aggregateSites}
\usage{
aggregateSites(
  m,
  ag = -5.5,
  b = 0.8656,
  c = -1.8488,
  p3 = NULL,
  bg = 0,
  toInt = FALSE,
  BP = NULL
)
}
\arguments{
\item{m}{A GRanges or data.frame of matches.}

\item{ag}{}

\item{b}{}

\item{c}{}

\item{p3}{}

\item{toInt}{Logical; whether to convert repression scores to integers}

\item{BP}{BPPARAM argument for multithreading}
}
\value{
a data.frame
}
\description{
aggregateSites
}
