% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{characterizeSeedMatches}
\alias{characterizeSeedMatches}
\title{characterizeSeedMatches}
\usage{
characterizeSeedMatches(x, seed = NULL)
}
\arguments{
\item{x}{A factor or character vector of matches, or a data.frame or GRanges
containing a `sequence` column.}

\item{seed}{The seed which sequences of `x` should match, or a 'KdModel' object.}
}
\value{
A data.frame, or a `GRanges` if `x` is a `GRanges`
}
\description{
characterizeSeedMatches
}
\examples{
characterizeSeedMatches(c("UAAACCACCC","CGAACCACUG"), "AAACCAC")
}
