% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressedKdModList.R
\name{compressKdModList}
\alias{compressKdModList}
\title{compressKdModList}
\usage{
compressKdModList(mods)
}
\arguments{
\item{mods}{A named list of objects of class `KdModel`}
}
\value{
A `CompressedKdModelList`
}
\description{
compressKdModList
}
