% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModelList.R
\name{conservation}
\alias{conservation}
\title{conservation}
\usage{
conservation(x)
}
\arguments{
\item{x}{A KdModelList, or a KdModel}
}
\value{
A vector of the conservation status for each miRNA
}
\description{
conservation
}
