% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compressedKdModList.R
\name{decompressKdModList}
\alias{decompressKdModList}
\title{decompressKdModList}
\usage{
decompressKdModList(mods)
}
\arguments{
\item{mods}{A `CompressedKdModelList`}
}
\value{
A list of `KdModels`
}
\description{
decompressKdModList
}
