% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{findSeedMatches}
\alias{findSeedMatches}
\title{findSeedMatches}
\usage{
findSeedMatches(
  seqs,
  seeds,
  shadow = 0L,
  onlyCanonical = FALSE,
  maxLogKd = c(-0.3, -0.3),
  keepMatchSeq = FALSE,
  minDist = 7L,
  p3.mismatch = TRUE,
  p3.maxLoop = 8L,
  p3.extra = FALSE,
  agg.params = .defaultAggParams(),
  ret = c("GRanges", "data.frame", "aggregated"),
  BP = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{seqs}{A character vector or `XStringSet` of sequences in which to look.}

\item{seeds}{A character vector of 7-nt seeds to look for. If RNA, will be 
reversed and complemented before matching. If DNA, they are assumed to be
the target sequence to look for. Alternatively, a list of objects of class
`KdModel` or an object of class `KdModelList` can be given.}

\item{shadow}{Integer giving the shadow, i.e. the number of nucleotides
hidden at the beginning of the sequence (default 0)}

\item{onlyCanonical}{Logical; whether to restrict the search only to canonical
binding sites.}

\item{maxLogKd}{Maximum log_kd value to keep (default 0). Set to Inf to disable.}

\item{keepMatchSeq}{Logical; whether to keep the sequence (including flanking
dinucleotides) for each seed match (default FALSE).}

\item{minDist}{Integer specifying the minimum distance between matches of the same 
miRNA (default 1). Closer matches will be reduced to the highest-affinity. To 
disable the removal of overlapping features, use `minDist=-Inf`.}

\item{p3.mismatch}{Logical; whether to allow mismatches in 3' alignment}

\item{p3.maxLoop}{The maximum loop size for the 3' alignment}

\item{agg.params}{a named list with slots `ag`, `b` and `c` indicating the 
parameters for the aggregation. Ignored if `ret!="aggregated"`.}

\item{ret}{The type of data to return, either "GRanges" (default), 
"data.frame" (lighter weight than GRanges), or "aggregated" (aggregated per 
transcript).}

\item{BP}{Pass `BiocParallel::MulticoreParam(ncores, progressbar=TRUE)` to enable 
multithreading.}

\item{verbose}{Logical; whether to print additional progress messages (default on if 
not multithreading)}

\item{seedtype}{Either RNA, DNA or 'auto' (default)}
}
\value{
A GRanges of all matches. If `seeds` is a `KdModel` or `KdModelList`, the 
`log_kd` column will report the ln(Kd) multiplied by 1000, rounded and saved as an 
integer.
}
\description{
findSeedMatches
}
\examples{
# we create mock RNA sequences and seeds:
seqs <- sapply(1:10, FUN=function(x) paste(sample(strsplit("ACGU", "")[[1]], 
                                     1000, replace=TRUE),collapse=""))
names(seqs) <- paste0("seq",1:length(seqs))
seeds <- c("AAACCAC", "AAACCUU")
findSeedMatches(seqs, seeds)
}
