% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{get3pAlignment}
\alias{get3pAlignment}
\title{get3pAlignment}
\usage{
get3pAlignment(
  seqs,
  mirseq,
  mir3p.start = 9L,
  allow.mismatch = TRUE,
  maxMirLoop = 5L,
  maxTargetLoop = 9L,
  maxLoopDiff = 4L,
  TGsub = TRUE
)
}
\arguments{
\item{seqs}{A set of sequences in which to look for 3' matches}

\item{mirseq}{The sequence of the miRNA}

\item{mir3p.start}{The position in `mirseq` in which to start looking}

\item{allow.mismatch}{Logical; whether to allow mismatches}

\item{maxMirLoop}{Maximum miRNA loop size}

\item{maxTargetLoop}{Maximum target loop size}

\item{maxLoopDiff}{Maximum size difference between miRNA and target loops}

\item{TGsub}{Logical; whether to allow T/G substitutions}
}
\value{
A data.frame with one row for each element of `seqs`, indicating the
size of the miRNA bulge, the size of the target mRNA bulge, the number of
mismatches at the 3' end, and the partial 3' alignment score (i.e. roughly
the number of consecutive matching nucleotides)
}
\description{
get3pAlignment
}
\examples{
get3pAlignment(seqs="NNAGTGTGCCATNN", mirseq="TGGAGTGTGACAATGGTGTTTG")
}
