% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get8merRange}
\alias{get8merRange}
\title{get8merRange}
\usage{
get8merRange(mod)
}
\arguments{
\item{mod}{A `KdModel`}
}
\value{
A numeric vector of length two
}
\description{
Returns the minimum and maximum 8-mer log-kd values
}
\examples{
data("SampleKdModel")
get8merRange(SampleKdModel)
}
