% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModels.R
\name{getKdModel}
\alias{getKdModel}
\title{getKdModel}
\usage{
getKdModel(kd, name = NULL)
}
\arguments{
\item{kd}{A data.frame with at least the columns "X12mer" and "log_kd", or the path to
such a data.frame}

\item{name}{The optional name of the miRNA}
}
\value{
A linear model of class `KdModel`
}
\description{
Summarizes the binding affinity of 12-mers using linear models.
}
