% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModels.R
\name{getKmers}
\alias{getKmers}
\title{getKmers}
\usage{
getKmers(n = 4, from = c("G", "C", "T", "A"))
}
\arguments{
\item{n}{Number of elements}

\item{from}{Letters sampled}
}
\value{
A character vector
}
\description{
Returns all combinations of `n` elements of `from`
}
\examples{
getKmers(3)
}
