% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModels.R
\name{getSeed8mers}
\alias{getSeed8mers}
\title{getSeed8mers}
\usage{
getSeed8mers(seed, addNs = FALSE)
}
\arguments{
\item{seed}{The miRNA seed (target DNA sequence), a character vector of length 8 (if 
of length 7, a "A" will be added on the right)}

\item{addNs}{Logical; whether to include 8mers with one flanking N}
}
\value{
A vector of 1024 8mers.
}
\description{
Generates all possible 8mers with 4 consecutive and positioned matches to a given
seed.
}
\examples{
head(getSeed8mers("ACACTCCA"))
}
