% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getTranscriptSequence}
\alias{getTranscriptSequence}
\title{getTranscriptSequence}
\usage{
getTranscriptSequence(
  tx,
  species = NULL,
  ensdbs = NULL,
  genome = NULL,
  UTRonly = TRUE,
  ...
)
}
\arguments{
\item{tx}{The ensembl ID of the transcript}

\item{species}{The species, either 'hsa', 'rno', or 'mmu'; ignored if 
`ensdbs` is given and is an `EnsDb` object.}

\item{ensdbs}{An \code{\link[ensembldb]{EnsDb}} object, or a named list of 
such objects.}

\item{genome}{The genome sequence (e.g. 
\code{\link[BSgenome]{BSgenome-class}}). If one of the three pre-defined 
`species`, will attempt to fetch the corresponding packages (if isntalled).}

\item{UTRonly}{Logical; whether to fetch only the UTR sequences.}

\item{...}{Passed to \code{\link{AnnotationHub}}}
}
\value{
A \link{\code{DNAStringSet}}.
}
\description{
Utility wrapper to extracts the sequence of a given transcript (UTR or 
CDS+UTR).
}
\examples{
# too long to run (needs fetching the annotation) :
# getTranscriptSequence("ENST00000641515", species="hsa")
}
