% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModels.R
\docType{class}
\name{KdModel}
\alias{KdModel}
\alias{KdModel-methods}
\alias{KdModel-class}
\alias{show,KdModel-method}
\alias{summary,KdModel-method}
\alias{c,KdModel-method}
\title{miRNA affinity models}
\usage{
\S4method{show}{KdModel}(object)

\S4method{summary}{KdModel}(object)

\S4method{c}{KdModel}(x, ...)
}
\arguments{
\item{object, x, ...}{An object of class \code{\link{KdModel}}}
}
\value{
Depends on the method.
}
\description{
Methods for the \code{\link{KdModel}} class
}
\examples{
data(SampleKdModel)
SampleKdModel
summary(SampleKdModel)
}
\seealso{
\code{\link{KdModel}}, \code{\link{KdModelList}}
}
