% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModelList.R
\name{KdModelList-methods}
\alias{KdModelList-methods}
\alias{summary,KdModelList-method}
\alias{[,KdModelList,ANY-method}
\title{Methods for the \code{\link{KdModelList}} classes}
\usage{
\S4method{summary}{KdModelList}(object)

\S4method{[}{KdModelList,ANY}(x, i, j = NULL, ..., drop = TRUE)
}
\arguments{
\item{object, x}{An object of class \code{\link{KdModelList}}}

\item{i}{the index of item(s) to select}

\item{j, drop, ...}{ignored}
}
\value{
Depends on the method.
}
\description{
Methods for the \code{\link{KdModelList}} classes
}
\examples{
# create a KdModelList :
data(SampleKdModel)
kml <- KdModelList( SampleKdModel, SampleKdModel, makeUnique=TRUE )
summary(kml)
kml[1] # returns a KdModelList
kml[[2]] # returns a KdModel
conservation(kml)
}
\seealso{
\code{\link{KdModel}}, \code{\link{KdModelList}}
}
