% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedFst.R
\name{loadIndexedFst}
\alias{loadIndexedFst}
\title{loadIndexedFst}
\usage{
loadIndexedFst(file, nthreads = 1)
}
\arguments{
\item{file}{Path to the fst file, it's index (.idx), or their prefix.}

\item{nthreads}{Number of threads to use for reading (default 1). This does not affect
the loading of the index itself, but will affect all downstream reading operations
performed on the object.}
}
\value{
An object of class IndexedFst
}
\description{
Loads the index of an FST file, for fast random access.
}
\seealso{
\code{\link{IndexedFst-class}}
}
