% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotKdModel}
\alias{plotKdModel}
\title{plotKdModel}
\usage{
plotKdModel(mod, what = c("both", "seeds", "logo"), n = 10)
}
\arguments{
\item{mod}{A `KdModel`}

\item{what}{Either 'seeds', 'logo', or 'both' (default)}
}
\value{
A plot
}
\description{
plotKdModel
}
