% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{predictKD}
\alias{predictKD}
\title{predictKD}
\usage{
predictKD(kmer, mod)
}
\arguments{
\item{kmer}{The 12-mer sequences for which affinity should be predicted}

\item{mod}{A `KdModel`}
}
\value{
A vector of the same length as `kmer` indicating the corresponding
predicted affinities.
}
\description{
predictKD
}
