% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{removeOverlappingMatches}
\alias{removeOverlappingMatches}
\title{removeOverlappingMatches}
\usage{
removeOverlappingMatches(x, minDist = 7L, method = c("exact", "fast"))
}
\arguments{
\item{x}{A GRanges, sorted by (decreasing) importance}

\item{minDist}{Minimum distance between ranges}
}
\value{
A filtered GRanges.
}
\description{
Removes elements from a GRanges that overlap (or are within a given distance of) other 
elements higher up in the list (i.e. assumes that the ranges are sorted in order of
priority). The function handles overlaps between more than two ranges by successively
removing those that overlap higher-priority ones
}
