% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SampleKdModel}
\alias{SampleKdModel}
\title{Example KdModel (hsa-miR-155-5p)}
\value{
a `KdModel` object
}
\description{
`KdModel` for hsa-miR-155-5p, based on Kd predictions from the CNN of
[McGeary, Lin et al. (2019)](https://dx.doi.org/10.1126/science.aav1741).
}
