% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexedFst.R
\name{saveIndexedFst}
\alias{saveIndexedFst}
\title{saveIndexedFst}
\usage{
saveIndexedFst(d, index.by, file.prefix, nthreads = 1, ...)
}
\arguments{
\item{d}{A data.frame}

\item{index.by}{A column of `d` by which it should be indexed.}

\item{file.prefix}{Path and prefix of the output files.}

\item{nthreads}{Number of threads to use when writing (default 1). If NULL, will use
`threads_fst()`}

\item{...}{Passed to `write.fst`}
}
\value{
Nothing.
}
\description{
Saves a data.frame into an indexed FST file.
}
\seealso{
\code{\link{IndexedFst-class}}
}
