% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.server.R
\name{scanMiR.server}
\alias{scanMiR.server}
\title{scanMiR.server}
\usage{
scanMiR.server(
  modlists,
  targetlists = list(),
  ensdbs = list(),
  genomes = list(),
  scans = list(),
  maxCacheSize = 100 * 10^6
)
}
\arguments{
\item{modlists}{A named list of `CompressedKdModelList`}

\item{targetlists}{An optional list of aggregated targets, named as `modlists`}

\item{ensdbs}{A named list of `ensembldb` objects (with names corresponding to those 
of `modlists`)}

\item{genomes}{A named list of `BSgenome` objects (with names corresponding to those 
of `modlists`)}

\item{maxCacheSize}{Maximum cache size in bytes.}
}
\value{
A shiny server function
}
\description{
scanMiR.server
}
