% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{viewTargetAlignment}
\alias{viewTargetAlignment}
\title{viewTargetAlignment}
\usage{
viewTargetAlignment(m, miRNA)
}
\arguments{
\item{m}{A GRanges of length 1 giving the information for a given match, as
produced by \link{\code{findSeedMatches}}.}

\item{miRNA}{A miRNA sequence, or a \link{\code{KdModel}} object of the miRNA
corresponding to the match in `m`}
}
\value{
A data.frame containing the aligned sequences.
}
\description{
viewTargetAlignment
}
