% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregateSites}
\alias{aggregateSites}
\title{aggregateSites}
\usage{
aggregateSites(
  m,
  ag = -4.863126,
  b = 0.5735,
  c = -1.7091,
  p3 = 0.04403,
  coef_utr = -0.28019,
  coef_orf = -0.08622,
  p3.range = c(3L, 8L),
  keepSiteInfo = FALSE,
  toInt = FALSE,
  BP = NULL
)
}
\arguments{
\item{m}{A GRanges or data.frame of matches.}

\item{toInt}{Logical; whether to convert repression scores to integers}

\item{BP}{BPPARAM argument for multithreading}
}
\value{
a data.frame
}
\description{
aggregateSites
}
