% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{get3pAlignment}
\alias{get3pAlignment}
\title{get3pAlignment}
\usage{
get3pAlignment(
  seqs,
  mirseq,
  mir3p.start = 9L,
  allow.mismatch = TRUE,
  TGsub = TRUE
)
}
\arguments{
\item{seqs}{A set of sequences in which to look for matches}

\item{mirseq}{The sequence of the miRNA}

\item{mir3p.start}{The position in `mirseq` in which to start looking}

\item{allow.mismatch}{Logical; whether to allow mismatches}

\item{TGsub}{Logical; whether to allow T/G substitutions}
}
\value{
A data.frame with one row for each element of `seqs`.
}
\description{
get3pAlignment
}
\examples{
get3pAlignment(seqs="NNAGTGTGCCATNN", mirseq="TGGAGTGTGACAATGGTGTTTG")
}
