% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{get3pAlignment}
\alias{get3pAlignment}
\title{get3pAlignment}
\usage{
get3pAlignment(
  seqs,
  mirseq,
  mir3p.nts = 8L,
  mir3p.start = 12L,
  extra.3p = TRUE,
  subm = NULL
)
}
\arguments{
\item{seqs}{A set of sequences in which to look for matches}

\item{mirseq}{The sequence of the miRNA}

\item{mir3p.nts}{The number of miRNA nucelotide in which to look 
for matches}

\item{mir3p.start}{The position in `mirseq` in which to start looking}

\item{extra.3p}{Logical; whether to provide more detail information about the
alignment.}

\item{subm}{An optional substitution matrix; by default a binary diagonal 
matrix with additional 0.65 on G/T is used.}
}
\value{
A data.frame with one row for each element of `seqs`.
}
\description{
get3pAlignment
}
\examples{
get3pAlignment(target="NNAGTGTGCCATNN", mirseq="TGGAGTGTGACAATGGTGTTTG")
}
