% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getRandomSeq}
\alias{getRandomSeq}
\title{getRandomSeq}
\usage{
getRandomSeq(length = 3000, alphabet = c("A", "C", "G", "T"), n = 1)
}
\arguments{
\item{length}{Length of the sequence}

\item{alphabet}{Letters from which to sample}

\item{n}{The number of sequences to generate}
}
\value{
A character vector of length 1
}
\description{
Produces a random sequence of the given letters
}
\examples{
getRandomSeq(100)
}
