% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KdModelList.R
\name{KdModelList}
\alias{KdModelList}
\title{KdModelList}
\usage{
KdModelList(x, description = NULL)
}
\arguments{
\item{x}{A list of KdModels}

\item{description}{A description for the collection.}
}
\value{
A KdModelList
}
\description{
KdModelList
}
\examples{
data(SampleKdModel)
KdModelList(list(SampleKdModel,SampleKdModel))
}
