% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotKdModel}
\alias{plotKdModel}
\title{plotKdModel}
\usage{
plotKdModel(mod, what = c("both", "seeds", "logo"), n = 10)
}
\arguments{
\item{mod}{A `KdModel`}

\item{what}{Either 'seeds', 'logo', or 'both' (default)}
}
\value{
If `what="logo"`, returns nothing a plots a position weight matrix.
Otherwise returns a ggplot.
}
\description{
Plots the summary of an affinity model. Requires the packages `seqLogo` and 
`cow_plot`.
}
