% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanning.R
\name{runFullScan}
\alias{runFullScan}
\title{runFullScan}
\usage{
runFullScan(
  species,
  mods = NULL,
  UTRonly = TRUE,
  shadow = 15,
  cores = 8,
  minLogKd = 0,
  save.path = NULL,
  ...
)
}
\description{
runFullScan
}
