% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{viewTargetAlignment}
\alias{viewTargetAlignment}
\title{viewTargetAlignment}
\usage{
viewTargetAlignment(
  m,
  miRNA,
  seqs = NULL,
  flagBulgeMatches = FALSE,
  maxBulgeSize = 9L,
  maxBulgeDiff = 4L,
  min3pMatch = 3L,
  UGsub = TRUE,
  ...,
  outputType = c("print", "data.frame", "plot", "ggplot")
)
}
\arguments{
\item{m}{A GRanges of length 1 giving the information for a given match, as
produced by \link{\code{findSeedMatches}}.}

\item{miRNA}{A miRNA sequence, or a \link{\code{KdModel}} object of the miRNA
corresponding to the match in `m`; alternatively, a \link{\code{KdModelList}}
including the model.}

\item{seqs}{The sequences corresponding to the seqnames of `m`. Not needed if
`m` contains the target sequences.}

\item{flagBulgeMatches}{Logical; whether to flag matches inside the bulge
(default FALSE)}

\item{maxBulgeSize}{The maximum bulge size to consider (default none)}

\item{maxBulgeDiff}{The maximum difference between miRNA and target bulges}

\item{min3pMatch}{The minimum 3' alignment for any to be plotted}

\item{UGsub}{Logical; whether to show U-G matches}

\item{...}{Passed to `text` if `outputType="plot"`.}

\item{outputType}{Either 'print' (default, prints to console), 'data.frame',
or 'plot'.}
}
\value{
Returns nothing `outputType="print"`. If `outputType="data.frame"`, 
returns a data.frame containing the alignment strings; if 
`outputType="plot"` returns a `ggplot` object.
}
\description{
viewTargetAlignment
}
\examples{
data(SampleKdModel)
seq <- c(seq1="CGACCCCTATCACGTCCGCAGCATTAAAT")
m <- findSeedMatches(seq, SampleKdModel)
viewTargetAlignment(m, miRNA=SampleKdModel, seqs=seq)
}
