% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{viewTargetAlignment}
\alias{viewTargetAlignment}
\title{viewTargetAlignment}
\usage{
viewTargetAlignment(
  m,
  miRNA,
  seqs = NULL,
  flagBulgeMatches = FALSE,
  maxBulgeSize = Inf,
  maxBulgeDiff = Inf
)
}
\arguments{
\item{m}{A GRanges of length 1 giving the information for a given match, as
produced by \link{\code{findSeedMatches}}.}

\item{miRNA}{A miRNA sequence, or a \link{\code{KdModel}} object of the miRNA
corresponding to the match in `m`; alternatively, a \link{\code{KdModelList}}
including the model.}

\item{seqs}{The sequences corresponding to the seqnames of `m`. Not needed if
`m` contains the target sequences.}

\item{flagBulgeMatches}{Logical; whether to flag matches inside the bulge
(default FALSE)}

\item{maxBulgeSize}{The maximum bulge size to consider (default none)}

\item{maxBulgeDiff}{The maximum difference between miRNA and target bulges}
}
\value{
A data.frame containing the aligned sequences.
}
\description{
viewTargetAlignment
}
